# openapi.yml
openapi: '3.0.2'
info:
  title: Clique Ledger RESTAPI
  version: '1.0'
servers:
  - url: https://api.cliquesledger.com/v1
...
paths:
...
  /transactions:
    get:
      tags:
        - Transactions
      parameters:
        - in: query
          name: sender
          description: This contains sender id
          schema:
            type: string
            example: U00002
        - in: query
          name: receiver
          description: This contains receiver Id
          schema:
            type: string
            example: U00002
        - in: query
          name: cliqueId
          description: This contains cliques Id
          schema:
            type: string
            example: L000001
        - in: query
          name: from-date
          description: From date of trasaction
          schema:
            type: string
            example: 2022-06-12
        - in: query
          name: to-date
          description: To date of transaction
          schema:
            type: string
            example: 2022-08-12
        - in: query
          name: limit
          description: Pagination limit
          schema:
            type: integer
            example: 10
        - in: query
          name: offset
          description: Pagination offset
          schema:
            type: integer
            example: 2

          
      responses:
        200:
          description: To get all transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: "#/components/schemas/TransactionResponse"
                    - $ref: "#/components/schemas/TransactionResponse"
                  
    post:
      tags:
        - Transactions
      requestBody:
        content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionRequest"
          
      responses:
        201:
          description: To get all transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
                status: SUCCESS
  /transactions/{transactionId}:
      get:
        tags: 
          - Transactions
        parameters:
          - $ref: '#/components/parameters/TransactionIdParam'
        responses:
          200:
            description: To get all transaction for a specific Id
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/TransactionResponse'

      delete:
        tags:
          - Transactions
        parameters:
          - in: path
            name: transactionId
            required: true
            schema:
              type: string
            description: The ID of the transaction to delete.
        responses:
          '204':
            description: Transaction deleted successfully.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: Transaction deleted successfully
          '404':
            description: Transaction not found.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    error:
                      type: string
                      example: Transaction not found
          '500':
            description: Internal server error.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    error:
                      type: string
                      example: An error occurred while deleting the transaction
        security:
          - bearerAuth: []
  /transaction/{transactionId}/verify/accept:
    patch:
      tags: 
        - Transactions
      parameters:
        - $ref: '#/components/parameters/TransactionIdParam'
      responses:
        200:
          description: It should work only for send trasaction
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: 
                    type: string
                    example: SUCCESS
                  message: 
                    type: string
                    example: successfully verified transaction
        400:
          description: In case someone tries to accept any other trasaction type appart from send
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:   
                    type: string
                    example: FAILED
                  message: 
                    type: string
                    example: This type of transaction does not allow such action
  /transaction/{transactionId}/verify/reject:
    patch:
      tags: 
        - Transactions
      parameters:
        - $ref: '#/components/parameters/TransactionIdParam'
      responses:
        200:
          description: It should work only for send trasaction
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: 
                    type: string
                    example: SUCCESS
                  message: 
                    type: string
                    example: successfully rejected transaction
        400:
          description: In case someone tries to reject any other trasaction type appart from send
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:   
                    type: string
                    example: FAILED
                  message: 
                    type: string
                    example: This type of transaction does not allow such action
  /transactions/{transactionId}/participants:
    post:
      tags:
        - Transactions
      parameters:
        - $ref: '#/components/parameters/TransactionIdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              example: ['M0005', 'M00010']

      responses:
        204:
          description: When the memeber is added successfully
        400:
          description: In case participant list contains wrong id. None of theme will be added
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: 
                    type: string
                    example: FAILED
                  message:
                    type: string
                    example: One or more participant id is provided wrong
        403:
          description: When non admin tries to add someone
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: 
                    type: string
                    example: FAILED
                  message: 
                    type: string
                    example: Permission Denined! Only admins can do this action.
        404:
          description: When transactionId is provided wrong
          content:
            application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: string
                      example: FAILED
                    message:
                      type: string
                      example: transactionId is provided wrong
    delete:
      tags: 
        - Transactions
      parameters:
        - $ref: '#/components/parameters/TransactionIdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              example: ['M00010']
      responses:
        204:
          description: When it is deleted successfully
        400:
          description: In case participant list contains wrong id. None of theme will be added
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: 
                    type: string
                    example: FAILED
                  message:
                    type: string
                    example: One or more participant id is provided wrong
        403:
          description: When non admin tries to add someone
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: 
                    type: string
                    example: FAILED
                  message: 
                    type: string
                    example: Permission Denined! Only admins can do this action.
        404:
          description: When transactionId is provided wrong
          content:
            application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: string
                      example: FAILED
                    message:
                      type: string
                      example: transactionId is provided wrong
...
components:
...
  parameters:
...
    TransactionIdParam:
        in: path
        name: transactionId
        description: This contains member Id
        required: true
        schema:
          type: string
          example: TN000001
  schemas:
...
    TransactionRequest:
      type: object
      properties: 
          type:
            type: string
            example: spend
          cliqueId:
            type: string
            example: C000001
          sender:
            type: string
            example: M00002
          paritcipants:
            type: array
            items:
              type: object
              properties:
                id: 
                  type: string
                  example: M00002
                amount: 
                  type: number
                  example: 300
          amount:
            type: number
            example: 1000
    TransactionResponse:
      type: object
      properties:
        transaction_id:
          type: string
          example: TN00001
        clique_id:
          type: string
          example: C000001
        transaction_type:
          type: string
          example: spend
        sender:
          type: object
          properties:
            member_id:
              type: string
              example: M000001
            member_name:
              type: string
              example: Sender Name
        is_verified:
          type: string
          example: "not_verified"
        done_at:
          type: string
          format: date-time
          example: "2023-07-04T12:34:56Z"
        participants:
          type: array
          items:
            type: object
            properties:
              member_id:
                type: string
                example: M000002
              member_name:
                type: string
                example: Participant Name
              part_amount:
                type: number
                example: 300
        amount:
          type: number
          example: 1000
        description:
          type: string
          example: "Dinner with friends"
...